requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
bcrypt>=3.2.0
alembic>=1.7.3
pytest>=6.2.5
httpx>=0.19.0
cryptography>=36.0.0  # For secure token generation and encryption
email-validator
pydantic-settings